pyproject.toml
[project]
name = "fast-rail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-fasthtml>=0.12.24",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
